rename code_gen_module -> code_generator
