+++ atcodertools/codegen/code_style_config.py
@@ -14,7 +14,7 @@ class CodeStyleConfig:
     def __init__(self,
                  indent_type: str = INDENT_TYPE_SPACE,
                  indent_width: int = 4,
-                 code_gen_module_file: str = None,
+                 code_generator_file: str = None,
                  ):
 
         if indent_type not in [INDENT_TYPE_SPACE, INDENT_TYPE_TAB]:
@@ -25,22 +25,22 @@ class CodeStyleConfig:
             raise CodeStyleConfigInitError(
                 "indent_width must be a positive integer")
 
-        if code_gen_module_file is not None and not os.path.exists(code_gen_module_file):
+        if code_generator_file is not None and not os.path.exists(code_generator_file):
             raise CodeStyleConfigInitError(
-                "Module file {} is not found".format(code_gen_module_file))
+                "Module file {} is not found".format(code_generator_file))
 
         self.indent_type = indent_type
         self.indent_width = indent_width
-        self.code_gen_module = None
+        self.code_generator = None
 
-        if code_gen_module_file is not None:
+        if code_generator_file is not None:
             try:
                 module = imm.SourceFileLoader(
-                    'code_gen_module', code_gen_module_file).load_module()
-                self.code_gen_module = getattr(module, 'main')
+                    'code_generator', code_generator_file).load_module()
+                self.code_generator = getattr(module, 'main')
             except AttributeError as e:
                 raise CodeStyleConfigInitError(e, "Error while loading {}".format(
-                    code_gen_module_file))
+                    code_generator_file))
 
     def indent(self, depth):
         if self.indent_type == INDENT_TYPE_SPACE:
