+++ atcodertools/tools/atcodertools-default.toml
@@ -2,9 +2,9 @@
 indent_type = 'space' # 'tab' or 'space'
 indent_width = 4
 
-## code_gen_module_file is a path to .py file defining custom template parameters generator function.
+## code_generator_file is a path to .py file defining custom template parameters generator function.
 ## Please check README.md for details.
-#code_gen_module_file=
+#code_generator_file=
 
 [postprocess]
 ## exec_on_each_problem_dir is executed after every problem directory's creation
