+++ tests/resources/test_config/broken_custom_code_generator.py
