+++ tests/test_codegen.py
@@ -3,7 +3,7 @@ import sys
 import tempfile
 import unittest
 
-from atcodertools.codegen.code_gen_modules import cpp, java
+from atcodertools.codegen.code_generators import cpp, java
 from atcodertools.codegen.code_style_config import CodeStyleConfig
 from atcodertools.codegen.models.code_gen_args import CodeGenArgs
 from atcodertools.codegen.template_engine import render
